Admin login copy chris

+++ NodeJs/myapp/routes/categories.js
@@ -2,13 +2,14 @@ const express = require("express")
 const categoriesModel = require('../models/categories')
 const categoriesRouter = express.Router();
 const categoriesController = require('../controllers/categories')
+const booksModel = require('../models/books');
 
 categoriesRouter.post("/", async (req, res) => {
 
     const categoriesInstance = new categoriesModel({
         name: req.body.name,
-        categoryId: req.body.categoryId,
-        books: req.body.books
+        // categoryId: req.body.categoryId,
+        //books: req.body?.books
     })
 
     const category = await categoriesInstance.save()
@@ -45,9 +46,6 @@ categoriesRouter.post("/", async (req, res) => {
         res.setHeader('Access-Control-Allow-Origin', 'http://localhost:4200');
         const category = await categoriesModel.find({}).populate("books").exec()
         try {
-           
-            console.log(category);
-           
             res.json(category)
         }
         catch (err) {
@@ -67,15 +65,21 @@ categoriesRouter.post("/", async (req, res) => {
     })
 
     .delete("/:id", async (request, response) => {
-        const { id } = request.params
         try {
-            const deletedcategory = await categoriesModel.deleteOne({ _id: id })
-            response.send("categoriesModel Deleted Correctly")
+            const { id } = request.params
+            const category = await categoriesModel.findById(id);
+            if (category.books.length < 1) {
+                const deletedcategory = await categoriesModel.deleteOne({ _id: id });
+                response.status(200).send({ deleted: true, message: "Category deleted succeffully" });
+            } else {
+                response.status(202).send({
+                    deleted: false, message: "This Category Contains books"
+                });
+            }
         } catch (error) {
-            return console.log(error);
+            response.status(400).send(error.message);
         }
     })
-
     .patch("/:id", async (request, response) => {
         const { id } = request.params;
         const category = request.body
@@ -84,10 +88,9 @@ categoriesRouter.post("/", async (req, res) => {
             ...(category.categoryId ? { categoryId: category.categoryId } : {})
 
         }
-
         try {
             const updatededcategory = await categoriesModel.findOneAndUpdate({ _id: id }, updatedcategory)
-            response.json(updatededcategory)
+            response.status(200).send({ message: "category edited succeffully" })
         } catch (error) {
             return console.log(error);
         }
